LayoutTests/http/tests/websocket/tests/hybi/url-with-credential-expected.txt

Make sure handshake with URL with credentail info success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://user:pass@127.0.0.1:8880/websocket/tests/hybi/echo-location
WebSocket is open
received:ws://127.0.0.1:8880/websocket/tests/hybi/echo-location
WebSocket is closed
PASS handshake_success is true
PASS ws_location is "ws://127.0.0.1:8880/websocket/tests/hybi/echo-location"
PASS successfullyParsed is true

TEST COMPLETE
